train/requirements.txt
python-dotenv==0.20.0
tensorflow==2.9.1
numpy==1.22.3
minio==7.1.8